Move output tests to separate files, use testsets
